feat: combine logic from both setups
