+++ test/a11y/axe.next.test.ts
@@ -1,10 +1,8 @@
 import { expect, test, type Page } from '@playwright/test';
-import AxeBuilder from '@axe-core/playwright';
 import * as cheerio from 'cheerio';
 import { AxeResults } from 'axe-core';
 import { readFileSync } from 'fs';
-import { writeFile } from 'fs/promises';
-import { exclusions, exclusionGroups } from './a11y-exclusions';
+import { analyzeAccessibility, getDisabledRules, saveViolationsReport } from './test-setup';
 
 const CONFIG = {
   baseUrl: 'http://localhost:4321',
@@ -51,7 +49,7 @@ function getPathnamesFromSitemap(sitemapPath: string): string[] {
       .map((url) => new URL(url).pathname)
       .forEach((path) => paths.push(path));
 
-    return paths;
+    return Array.from(new Set(paths));
   } catch (error) {
     console.warn(`Could not read sitemap at ${sitemapPath}, skipping accessibility tests generation.`, error);
     return [];
@@ -69,49 +67,7 @@ async function verifyPageAccessibility(page: Page, pathname: string): Promise<vo
   const results = await analyzeAccessibility(page, disabledRules);
 
   violations.push(results);
-  expect(results.violations).toEqual([]);
-}
 
-async function analyzeAccessibility(page: Page, disabledRules: string[]) {
-  return new AxeBuilder({ page })
-    .options({ resultTypes: ['violations'] })
-    .withTags(['wcag2a', 'wcag2aa', 'wcag21a', 'wcag21aa', 'wcag22aa'])
-    .disableRules(disabledRules)
-    .analyze();
-}
-
-async function saveViolationsReport(reportData: unknown[], filePath: string): Promise<void> {
-  console.log(`Writing accessibility report to ${filePath}`);
-  await writeFile(filePath, JSON.stringify(reportData, null, 2));
-}
-
-function getDisabledRules(pathname: string): string[] {
-  const disabledRules = new Set<string>();
-
-  for (const exclusion of exclusions) {
-    const isMatch = exclusion.routes.some((route) => {
-      if (typeof route === 'string') {
-        return route === pathname;
-      }
-      return route.test(pathname);
-    });
-
-    if (isMatch) {
-      if (exclusion.rules) {
-        exclusion.rules.forEach((rule) => disabledRules.add(rule));
-      }
-      if (exclusion.groups) {
-        exclusion.groups.forEach((groupName) => {
-          const groupRules = exclusionGroups[groupName];
-          if (groupRules) {
-            groupRules.forEach((rule) => disabledRules.add(rule));
-          } else {
-            console.warn(`Warning: Exclusion group '${groupName}' not found in configuration.`);
-          }
-        });
-      }
-    }
-  }
-
-  return Array.from(disabledRules);
+  const criticalViolations = results.violations.filter((v) => v.impact === 'critical');
+  expect(criticalViolations).toEqual([]);
 }
