+++ test/a11y/axe.test.ts
@@ -1,10 +1,8 @@
 import { expect, test, type Page } from '@playwright/test';
-import AxeBuilder from '@axe-core/playwright';
 import * as cheerio from 'cheerio';
 import { AxeResults } from 'axe-core';
 import { readFileSync } from 'fs';
-import { writeFile } from 'fs/promises';
-import { exclusions, exclusionGroups, skippedRoutes } from './a11y-exclusions';
+import { analyzeAccessibility, getDisabledRules, saveViolationsReport, shouldSkipRoute } from './test-setup';
 
 const CONFIG = {
   baseUrl: 'http://localhost:3000',
@@ -33,10 +31,12 @@ function getPathnamesFromSitemap(sitemapPath: string): string[] {
     const sitemap = readFileSync(sitemapPath, 'utf8');
     const $ = cheerio.load(sitemap, { xmlMode: true });
 
-    return $('loc')
+    const paths = $('loc')
       .map((_, element) => $(element).text())
       .toArray()
       .map((url) => new URL(url).pathname);
+
+    return Array.from(new Set(paths));
   } catch (error) {
     console.warn(`Could not read sitemap at ${sitemapPath}, skipping accessibility tests generation.`, error);
     return [];
@@ -55,65 +55,10 @@ async function verifyPageAccessibility(page: Page, pathname: string): Promise<vo
   const results = await analyzeAccessibility(page, disabledRules);
 
   violations.push(results);
-  expect(results.violations).toEqual([]);
+  const criticalViolations = results.violations.filter((v) => v.impact === 'critical');
+  expect(criticalViolations).toEqual([]);
 }
 
 function isDocusaurusHydrated(): boolean {
   return document.documentElement.dataset.hasHydrated === 'true';
-}
-
-function shouldSkipRoute(pathname: string): boolean {
-  return skippedRoutes.some((route) => {
-    if (typeof route === 'string') {
-      if (route.endsWith('*')) {
-        return pathname.startsWith(route.slice(0, -1));
-      }
-      return route === pathname;
-    }
-    return route.test(pathname);
-  });
-}
-
-async function analyzeAccessibility(page: Page, disabledRules: string[]) {
-  return new AxeBuilder({ page })
-    .options({ resultTypes: ['violations'] })
-    .withTags(['wcag2a', 'wcag2aa', 'wcag21a', 'wcag21aa', 'wcag22aa'])
-    .disableRules(disabledRules)
-    .analyze();
-}
-
-async function saveViolationsReport(reportData: unknown[], filePath: string): Promise<void> {
-  console.log(`Writing accessibility report to ${filePath}`);
-  await writeFile(filePath, JSON.stringify(reportData, null, 2));
-}
-
-function getDisabledRules(pathname: string): string[] {
-  const disabledRules = new Set<string>();
-
-  for (const exclusion of exclusions) {
-    const isMatch = exclusion.routes.some((route) => {
-      if (typeof route === 'string') {
-        return route === pathname;
-      }
-      return route.test(pathname);
-    });
-
-    if (isMatch) {
-      if (exclusion.rules) {
-        exclusion.rules.forEach((rule) => disabledRules.add(rule));
-      }
-      if (exclusion.groups) {
-        exclusion.groups.forEach((groupName) => {
-          const groupRules = exclusionGroups[groupName];
-          if (groupRules) {
-            groupRules.forEach((rule) => disabledRules.add(rule));
-          } else {
-            console.warn(`Warning: Exclusion group '${groupName}' not found in configuration.`);
-          }
-        });
-      }
-    }
-  }
-
-  return Array.from(disabledRules);
 }
